Challenge - Part 1
Start the creation of the project in VS Code, Creation of the Order_Data.py class, Creation of the resource file, where all the actions that the Order.robot is going to execute with Gherkin are made, Install the gherkin2robotframework package. Start of the creation of the orders, definition of the objects, definition of the locators, paths, etc.